% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leanpub_collaborator.R
\name{add_leanpub_collaborator}
\alias{add_leanpub_collaborator}
\title{Add Leanpub as Collaborator}
\usage{
add_leanpub_collaborator(course_dir = ".", auth_token = NULL)
}
\arguments{
\item{course_dir}{directory with course materials or a \code{course_check}
object}

\item{auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. If \code{NULL}, will use the logic
described in \code{gh::gh_whoami()} to look for a token stored in an environment
variable. Use \code{\link[usethis:usethis-defunct]{usethis::browse_github_pat()}} to help set up your PAT.}
}
\value{
Answer from the API as a \code{gh_response} object,
which is also a \code{list.} Failed requests will generate an \code{R} error.
}
\description{
Add Leanpub as Collaborator
}
