% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_mooc_app.R
\name{authorize_mooc_app}
\alias{authorize_mooc_app}
\title{Authorize Application for MOOC}
\usage{
authorize_mooc_app(cache = FALSE, use_oob = FALSE)
}
\arguments{
\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}}

\item{use_oob}{use a local webserver for the OAuth dance}
}
\value{
A list with the token and the saved file with the token
}
\description{
Authorize Application for MOOC
}
\examples{
\dontrun{
authorize_mooc_app()
}
}
