% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_course.R
\name{check_course}
\alias{check_course}
\title{Check Leanpub Course}
\usage{
check_course(course_dir = ".", course_title, save_metrics = TRUE,
  timezone = "America/New_York", ...)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{save_metrics}{Should an `rds` file be saved of the `data.frame`?}

\item{timezone}{Timezone to be used?}

\item{...}{arguments to pass to \code{\link{didactr_auth}}}
}
\value{
A data frame of the checked course.
}
\description{
Check Leanpub Course
}
