% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_videos.R
\name{create_videos}
\alias{create_videos}
\title{Create video from script and PNGs}
\usage{
create_videos(course_status = NULL)
}
\arguments{
\item{course_status}{output from \code{\link{check_course}}}
}
\value{
created mp4s
}
\description{
Create video from script and PNGs
}
