% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_code_chunks.R
\name{flag_code_chunks}
\alias{flag_code_chunks}
\title{Flag Code Chunks}
\usage{
flag_code_chunks(x)
}
\arguments{
\item{x}{A string of text with code chunks}
}
\value{
A \code{data.frame} with the original text and the flagged code
and the code removed
}
\description{
Flag Code Chunks
}
