% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ari.R
\name{gs_ari_upload}
\alias{gs_ari_upload}
\title{Convert Google Slides to Video and Upload to Youtube}
\usage{
gs_ari_upload(id, ...,
  ffmpeg_opts = "-vf \\"scale=trunc(iw/2)*2:trunc(ih/2)*2\\"",
  snippet = NULL, status = NULL, open_url = TRUE)
}
\arguments{
\item{id}{Identifier of google slides presentation, passed to
\code{\link{gs_convert}}}

\item{...}{Arguments passed to \code{\link{gs_ari}}}

\item{ffmpeg_opts}{options passed to \code{\link{ari_spin}}}

\item{snippet}{snippet list passed to \code{\link{upload_video}}}

\item{status}{status list passed to \code{\link{upload_video}}}

\item{open_url}{Should YouTube URL be opened?  Passed to
\code{\link{upload_video}}}
}
\value{
A list of results from \code{\link{gs_ari}} and
\code{\link{upload_video}}
}
\description{
Convert Google Slides to Video and Upload to Youtube
}
