% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_folder_df.R
\name{gs_folder_df}
\alias{gs_folder_df}
\title{Get Google Folder Contents IDs}
\usage{
gs_folder_df(id, slides_only = FALSE)
}
\arguments{
\item{id}{Folder id passed to \code{\link{drive_ls}} after passing
through \code{\link{as_id}}}

\item{slides_only}{Return only elements that are
presentations}
}
\value{
A \code{data.frame} of the identifiers and properties of the
objects
}
\description{
Get Google Folder Contents IDs
}
\examples{
\dontrun{
folder_id = "1GCBE9EAgY3-i1bq9UYf5ec5s9X5TOwEK"
df = gs_folder_df(folder_id)
}
}
