% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_slide_code.R
\name{gs_slide_code}
\alias{gs_slide_code}
\title{Get Slide Code from Google Slides}
\usage{
gs_slide_code(id, open = FALSE)
}
\arguments{
\item{id}{ID of Google Slides}

\item{open}{Should the `Rmd` be opened to an editor}
}
\value{
The output temporary_file
}
\description{
Get Slide Code from Google Slides
}
\examples{
\dontrun{
if (check_didactr_auth) {
id = "1Tg-GTGnUPduOtZKYuMoelqUNZnUp3vvg_7TtpUPL7e8"
gs_slide_code(id, open = TRUE)
}
}
}
