% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_to_drive_pngs.R
\name{gs_to_drive_pngs}
\alias{gs_to_drive_pngs}
\title{Convert Google Slides to PNGs and Upload to Drive folder}
\usage{
gs_to_drive_pngs(
  id,
  folder_name = "leanpub_pngs",
  verbose = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{id}{ID or URL to Google slides presentation.}

\item{folder_name}{folder to upload to on Google Drive.  If no folder
exists, then one will be created.  Also a `dribble` can be passed}

\item{verbose}{print diagnostic messages, higher numbers give more output}

\item{overwrite}{Overwrite the PNGs if they exist, passed to
\code{\link{drive_upload}}}
}
\value{
A list of results
}
\description{
Convert Google Slides to PNGs and Upload to Drive folder
}
\examples{
\dontrun{
url = paste0("https://docs.google.com/presentation/d/",
"1ywZbtFacZK0UIsnt2g-sheC9du_rw_7XZ1FX4rRt27M",
"/export/png?",
"id=1ywZbtFacZK0UIsnt2g-sheC9du_rw_7XZ1FX4rRt27M&pageid=g325fd519ca_0_5")

out = gs_to_drive_pngs(url)
}
}
