% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ffmpeg.R
\name{install_ffmpeg}
\alias{install_ffmpeg}
\title{Install \code{ffmpeg} from a command line installer}
\usage{
install_ffmpeg(force = FALSE)
}
\arguments{
\item{force}{Should \code{ffmpeg} be installed
if found in the path already?}
}
\value{
Either \code{NA} if not found after install or the path
to the installation.
}
\description{
Install \code{ffmpeg} from a command line installer
}
\note{
The default installer for Linux is \code{apt-get},
which may not be appropriate for your distribution and is not run
using \code{sudo}
}
\examples{
install_ffmpeg()
}
