% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_language_auth.R
\name{is_language_auth}
\alias{is_language_auth}
\title{Is Google Language API Authorized}
\usage{
is_language_auth()
}
\value{
A logical
}
\description{
Is Google Language API Authorized
}
\examples{
is_language_auth()
}
