% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanpub_render.R
\name{leanpub_render}
\alias{leanpub_render}
\title{Render Leanpub document}
\usage{
leanpub_render(md_file, ...)
}
\arguments{
\item{md_file}{Path to markdown file}

\item{...}{additional arguments to send to \code{\link{render}}
other than \code{input}, `output_dir`, and `output_format`}
}
\value{
HTML output filename
}
\description{
Render Leanpub document
}
