% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_course.R
\name{make_course}
\alias{make_course}
\title{Check for expected file structure and create missing}
\usage{
make_course(course_dir = NULL)
}
\arguments{
\item{course_dir}{directory with course materials}
}
\value{
A data frame of the checked course.
}
\description{
Check for expected file structure and create missing
}
