% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_locales.R
\name{ms_locales}
\alias{ms_locales}
\alias{ms_locale_df}
\alias{ms_languages}
\alias{ms_language_codes}
\title{Locales for Microsoft Translate}
\usage{
ms_locales()

ms_locale_df()

ms_languages()

ms_language_codes()
}
\value{
A list of Locales and the Speech Voices
}
\description{
Locales for Microsoft Translate
}
\examples{
ms_locales()
}
