% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_pdf_pages.R
\name{n_pdf_pages}
\alias{n_pdf_pages}
\title{Count number of pages in PDF}
\usage{
n_pdf_pages(file)
}
\arguments{
\item{file}{input from \code{\link{check_course}}}
}
\value{
number of pages in pdf
}
\description{
Count number of pages in PDF
}
