% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_images.R
\name{pdf_to_images}
\alias{pdf_to_images}
\title{Convert PDF to Other Images using ImageMagick}
\usage{
pdf_to_images(pdf_file, out_dir = NULL,
  extra.opts = "-density 300 -quality 100", stub = "img_\%04d",
  output_type = "png", convert = "convert")
}
\arguments{
\item{pdf_file}{Filename of PDF file}

\item{out_dir}{Output directory.  If `NULL`, then a temporary directory
will be used.  If files of the same name are in the `out_dir`, these
will be overwritten}

\item{extra.opts}{Extra options passed to \code{\link{im.convert}}}

\item{stub}{prefix of the output files, can include C-style printing of
digits and such}

\item{output_type}{Type of output, such as `png` or `jpg`}

\item{convert}{converter passed to \code{\link{im.convert}}}
}
\value{
A character vector of filenames
}
\description{
Convert PDF to Other Images using ImageMagick
}
\examples{
\dontrun{
ex_file = system.file("extdata", "example.pdf", package = "didactr")
res = pdf_to_images(ex_file)
}
}
