% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_convert.R
\name{rename_pngs_to_gs_ids}
\alias{rename_pngs_to_gs_ids}
\title{Rename set of PNGs to Google Slide IDs}
\usage{
rename_pngs_to_gs_ids(id, pngs)
}
\arguments{
\item{id}{File id passed to \code{\link{drive_download}} after passing
through \code{\link{as_id}}}

\item{pngs}{Vector of PNG filenames}
}
\value{
A \code{data.frame}, similar to output of \code{\link{gs_slide_df}}
}
\description{
Rename set of PNGs to Google Slide IDs
}
