% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_to_slide_notes.R
\name{script_to_slide_notes}
\alias{script_to_slide_notes}
\title{Upload script to Google Slide Deck Notes}
\usage{
script_to_slide_notes(course_dir = ".", ...)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{...}{arguments to pass to \code{\link{check_course}}}
}
\value{
A \code{data.frame} the same format as
a course summary
}
\description{
Upload script to Google Slide Deck Notes
}
