% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_course.R
\name{start_course}
\alias{start_course}
\title{Starting a Course, Light wrapper for \code{\link{make_course}}}
\usage{
start_course(course_name, root_path = ".")
}
\arguments{
\item{course_name}{Name of the course.  Any spaces will be set to
underscores}

\item{root_path}{The directory to put the course into.}
}
\value{
The output of \code{\link{make_course}}.
}
\description{
Starting a Course, Light wrapper for \code{\link{make_course}}
}
\examples{
root_path = tempfile()
course_name = "test this out"
sc = start_course(course_name, root_path)
}
