% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_manuscript.R
\name{strip_manuscript}
\alias{strip_manuscript}
\title{Strip Markdown from Manuscript}
\usage{
strip_manuscript(file)
}
\arguments{
\item{file}{A markdown file for parsing}
}
\value{
A \code{data.frame} with text and edits
}
\description{
Strip Markdown from Manuscript
}
\examples{
library(dplyr)
if (requireNamespace("english", quietly = TRUE)) {
 library(english)
 file = system.file("extdata", "08_DST_R_packages.md",
 package = "didactr")
 # from https://stackoverflow.com/questions/37462126/regex-match-markdown-link
 # md_regex = "(?:__|[*#])|\\[(.*?)\\]\\(.*?\\)"
 res = strip_manuscript(file)
 res
}
}
