% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_course_check.R
\name{summary.course_check}
\alias{summary.course_check}
\title{Summary of course check}
\usage{
\method{summary}{course_check}(object, ...)
}
\arguments{
\item{object}{An object of class \code{course_check}}

\item{...}{Not used}
}
\value{
A list of data frames of various checks
}
\description{
Summary of course check
}
\examples{
root_path = tempfile()
course_name = "test this out2"
book_txt =  system.file("extdata", "Book.txt", package = "didactr")
sc = create_course(course_name, root_path, book_txt = book_txt)
course_dir = sc$course_dir
object = check_course(sc$course_dir,
require_authorization = FALSE)
summary(object)
}
