% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_manuscript_backticks.R,
%   R/translate_manuscript.R
\name{check_manuscript_backticks}
\alias{check_manuscript_backticks}
\alias{translate_manuscript}
\alias{translate_script}
\alias{chunk_google_translate}
\alias{gl_detect_file}
\title{Translate Mansucript File}
\usage{
check_manuscript_backticks(file)

translate_manuscript(file, target = "es", chunk = TRUE,
  fix_header = TRUE, verbose = TRUE, ...)

translate_script(file, target = "es", chunk = TRUE,
  fix_header = FALSE, verbose = TRUE, ...)

chunk_google_translate(file, chunk = TRUE, target = "es",
  fix_header = TRUE, ...)

gl_detect_file(file, max_nchar = 2000)
}
\arguments{
\item{file}{Manuscript markdown file}

\item{target}{target language, see \code{\link{gl_translate}}}

\item{chunk}{Should the calls to Google Translate be chunked?}

\item{fix_header}{should the header information be fixed,
such as `{ course-completeness: 100 }`.}

\item{verbose}{Print diagnostic messages}

\item{...}{additional arguments to pass to
\code{\link{gl_translate}}}

\item{max_nchar}{number of characters to use for detection.}
}
\value{
A `data.frame` of the text and translated text
}
\description{
Translate Mansucript File
}
\note{
The following changes of the translated text are
attempted:
Change `] (` without spaces for links
Keep links to be the link from original text
Replace asterisks and underscores for bolding
Keep back-ticked code as is
}
\examples{
auth_file = "~/Dropbox/Projects/cbds_translate/RClass-Translator.json"
if (file.exists(auth_file)) {
googleLanguageR::gl_auth(auth_file)
target = "es"
verbose = TRUE
 chunk = TRUE
file = system.file("extdata", "00_template.md", package = "didactr")
df = translate_manuscript(file)
}
}
