% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_slides.R
\name{translate_slide}
\alias{translate_slide}
\alias{copy_and_translate_slide}
\title{Translate Google Slide Text to Another Language}
\usage{
translate_slide(id, target = "es", detect = TRUE, verbose = TRUE, ...)

copy_and_translate_slide(
  id,
  gs_name = NULL,
  trash_same_gs_name = FALSE,
  share = FALSE,
  target = "es",
  detect = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{id}{Google Slide ID, usually from \code{\link{drive_ls}}}

\item{target}{language to translate to, see \code{\link{gl_translate}}}

\item{detect}{Should \code{\link{gl_detect_file}} be run to
check that the language is not currently the target?
Must be google Language authorized using \code{\link{gl_auth}}}

\item{verbose}{Print diagnostic messages}

\item{...}{arguments passed to
\code{\link{check_didactr_auth}}}

\item{gs_name}{Name of new Google slide deck}

\item{trash_same_gs_name}{Should other Google slide decks
with the same name be trashed before copying?  If not,
can fill up your drive.}

\item{share}{should the link sharing be turned on?}
}
\value{
A list of results from \code{\link{commit_to_slides}}
}
\description{
Translate Google Slide Text to Another Language
}
\note{
Copies are put in the \code{didactr_translations}
folder in your Google Drive.
}
\examples{
\dontrun{
library(googledrive)
check_didactr_auth()
# id = "147UQaZBB5RoTpzsNiqkqEr4N8fcHBMB6eNr_5IdksRk"
info = drive_find("datatables", n_max = 2000, type = "presentation")
info = info[grepl("^06" ,info$name), ]
if (nrow(info) > 0) {
info = info[1,]
xid = drive_cp(info, name = paste0("Translated ",
info$name))
id = xid
translated = translate_slide(id, target = "es")
drive_trash(id)
}

}
\dontrun{
check_didactr_auth()
id = "1Opt6lv7rRi7Kzb9bI0u3SWX1pSz1k7botaphTuFYgNs"
res = copy_and_translate_slide(id)
googledrive::drive_trash(res$id)
}
}
