% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_swirl.R
\name{translate_swirl}
\alias{translate_swirl}
\title{Translate Swirl YAML to another language}
\usage{
translate_swirl(
  file,
  outfile = tempfile(fileext = ".yml"),
  target = "es",
  detect = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{YAML file for Swirl lesson}

\item{outfile}{output YAML file for translated Swirl lesson}

\item{target}{language to translate to, see \code{\link{gl_translate}}}

\item{detect}{Should \code{\link{gl_detect_file}} be run to
check that the language is not currently the target?
Must be google Language authorized using \code{\link{gl_auth}}}

\item{verbose}{Print diagnostic messages}

\item{...}{additional options to pass to \code{\link{write_yaml}}}
}
\value{
A filename of the output file.  The attribute
 \code{target_language} is also given to the filename.
}
\description{
Translate Swirl YAML to another language
}
\examples{
\dontrun{
x = c("- Class: meta",
"  Course: Test Class", "  Lesson: Test lesson",
"  Author: Shannon Ellis", "", "- Class: text",
"  Output: This is an example with code `df`.")
tfile = tempfile(fileext = ".yml")
writeLines(x, tfile)
translate_swirl(tfile, outfile = tfile)
readLines(tfile)
}
}
