% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_youtube_link.R
\name{update_youtube_link}
\alias{update_youtube_link}
\title{Update Youtube Link}
\usage{
update_youtube_link(course_status = NULL)
}
\arguments{
\item{course_status}{output from \code{\link{check_course}}}
}
\value{
updated manuscript files
}
\description{
Update Youtube Link
}
