% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_course.R
\name{check_course}
\alias{check_course}
\title{Check Leanpub Course}
\usage{
check_course(course_dir = NULL)
}
\arguments{
\item{course_dir}{directory with course materials}
}
\value{
A data frame of the checked course.
}
\description{
Check Leanpub Course
}
