% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_course.R
\name{check_course}
\alias{check_course}
\title{Check Leanpub Course}
\usage{
check_course(
  course_dir = ".",
  save_metrics = TRUE,
  timezone = "America/New_York",
  require_authorization = TRUE,
  use_book = FALSE,
  check_youtube_links = TRUE,
  ...
)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{save_metrics}{Should an `rds` file be saved of the `data.frame`?}

\item{timezone}{Timezone to be used?}

\item{require_authorization}{Should you authorize for Google
Slides and such to check the modified time?}

\item{use_book}{Use \code{Book.txt} to keep only the
manuscript files that were specified.}

\item{check_youtube_links}{Should YouTube links be checked?  If starting a
course, set to \code{FALSE}}

\item{...}{arguments to pass to \code{\link{didactr_auth}}}
}
\value{
A list of data frames of the checked course. Will have
class \code{course_check}
}
\description{
Check Leanpub Course
}
\examples{
root_path = tempfile()
course_name = "test this out2"
book_txt =  system.file("extdata", "Book.txt", package = "didactr")
sc = create_course(course_name, root_path, book_txt = book_txt)
course_dir = sc$course_dir
in_ci <- function() {
nzchar(Sys.getenv("CI"))
}
if (!in_ci()) {
object = check_course(sc$course_dir, require_authorization = FALSE)
}
}
