% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Check Course File Structure}
\usage{
check_structure(course_dir = ".")
}
\arguments{
\item{course_dir}{directory with course materials}
}
\value{
A list of the paths necessary for the course
}
\description{
Check Course File Structure
}
