% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_course.R
\name{make_course}
\alias{make_course}
\alias{check_structure}
\title{Check for expected file structure and create missing}
\usage{
make_course(course_dir = ".", book_txt = NULL, verbose = TRUE)

check_structure(course_dir = ".", book_txt = NULL, verbose = TRUE)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{book_txt}{A text file with names of markdown
files and course names}

\item{verbose}{print diagnostic messages}
}
\value{
A list of the paths necessary for the course
}
\description{
Check for expected file structure and create missing
}
\examples{
course_dir = tempfile()
dir.create(course_dir)
make_course(course_dir)
}
