% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_images.R
\name{create_images}
\alias{create_images}
\title{Create PDF and PNGs from Google Slides}
\usage{
create_images(course_status = NULL)
}
\arguments{
\item{course_status}{output from \code{\link{check_course}}}
}
\value{
Downloaded PDF and PNGs.
}
\description{
Create PDF and PNGs from Google Slides
}
