% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_videos.R
\name{create_videos}
\alias{create_videos}
\title{Create video from script and PNGs}
\usage{
create_videos(course_status = ".", voice = "Joanna",
  ffmpeg_opts = "-minrate 16M -vf \\"scale=1200:720\\"",
  divisible_height = FALSE, video_bitrate = "16M",
  audio_codec = NULL, verbose = TRUE, ...)
}
\arguments{
\item{course_status}{output from \code{\link{check_course}}}

\item{voice}{Voice to pass to \code{\link{ari_spin}}}

\item{ffmpeg_opts}{ffmpeg options to pass to \code{\link{ari_spin}}}

\item{divisible_height}{should heights be divisible?
Advanced usage to pass to \code{\link{ari_spin}}}

\item{video_bitrate}{video bitrate to enforce.
Advanced usage to pass to \code{\link{ari_spin}}}

\item{audio_codec}{audio codec to pass to \code{\link{ari_spin}}}

\item{verbose}{Should diagnostic messages be printed}

\item{...}{additional arguments passed to \code{\link{ari_spin}}}
}
\value{
A `data.frame` output from \code{\link{check_course}}
}
\description{
Create video from script and PNGs
}
