% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_videos.R
\name{create_videos}
\alias{create_videos}
\title{Create video from script and PNGs}
\usage{
create_videos(course_status = NULL, voice = "Joanna", audio_codec = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{course_status}{output from \code{\link{check_course}}}

\item{voice}{Voice to pass to \code{\link{ari_spin}}}

\item{audio_codec}{audio codec to pass to \code{\link{ari_spin}}}

\item{verbose}{Should diagnostic messages be printed}

\item{...}{additional arguments passed to \code{\link{ari_spin}}}
}
\value{
A `data.frame` output from \code{\link{check_course}}
}
\description{
Create video from script and PNGs
}
