% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_mooc_app.R
\name{didactr_auth}
\alias{didactr_auth}
\title{Authorize Application for didactr}
\usage{
didactr_auth(cache = FALSE, use_oob = FALSE)
}
\arguments{
\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}}

\item{use_oob}{use a local webserver for the OAuth dance}
}
\value{
The auth token
}
\description{
Authorize Application for didactr
}
\examples{
\dontrun{
didactr_auth()
}
}
