% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_mooc_app.R
\name{didactr_auth}
\alias{didactr_auth}
\alias{check_didactr_auth}
\alias{didactr_token}
\alias{is_didactr_authorized}
\title{Authorize Application for didactr}
\usage{
didactr_auth(token_file = NULL, cache = FALSE, use_oob = FALSE)

check_didactr_auth(...)

didactr_token(...)

is_didactr_authorized()
}
\arguments{
\item{token_file}{If the \code{token} has been saved, use this file
to load the credentials.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}}

\item{use_oob}{use a local webserver for the OAuth dance}

\item{...}{Arguments passed to \code{\link{didactr_auth}}}
}
\value{
The auth token, a Token class.
}
\description{
Authorize Application for didactr
}
\examples{
\dontrun{
didactr_auth()
}
}
