% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ari.R
\name{gs_ari}
\alias{gs_ari}
\alias{gs_ari_spin}
\title{Convert Google Slides and notes to video with ari}
\usage{
gs_ari(id, output = NULL, verbose = TRUE, voice = "Joanna", ...)

gs_ari_spin(id, output = NULL, verbose = TRUE, voice = "Joanna", ...)
}
\arguments{
\item{id}{Identifier of google slides presentation, passed to
\code{\link{gs_convert}}}

\item{output}{Output MP4 filename, optional}

\item{verbose}{Should diagnostics messages be printed}

\item{voice}{The Amazon Polly voice you want to use. See
\code{\link[aws.polly]{list_voices}} for more
information about what voices are available.}

\item{...}{Arguments passed to \code{\link{ari_spin}}}
}
\value{
A list of the images, script, and MP4 with the result.
}
\description{
Convert Google Slides and notes to video with ari
}
