% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_ari.R
\name{gs_ari}
\alias{gs_ari}
\alias{gs_ari_spin}
\alias{gs_tts}
\title{Convert Google Slides and notes to video with ari}
\usage{
gs_ari(
  id,
  output = NULL,
  script = NULL,
  verbose = TRUE,
  voice = "Joanna",
  use_gs_pngs = TRUE,
  ...
)

gs_ari_spin(
  id,
  output = NULL,
  script = NULL,
  verbose = TRUE,
  voice = "Joanna",
  use_gs_pngs = TRUE,
  ...
)

gs_tts(id, verbose = TRUE, voice = "Joanna", use_gs_pngs = TRUE, ...)
}
\arguments{
\item{id}{Identifier of google slides presentation, passed to
\code{\link{gs_convert}}}

\item{output}{Output MP4 filename, optional}

\item{script}{A vector strings that will be spoken over slides,
or a filename of the text.  Empty string are removed, please use
\code{;} if no words to be spoken.}

\item{verbose}{Should diagnostics messages be printed}

\item{voice}{The Amazon Polly voice you want to use. See
\code{\link[aws.polly]{list_voices}} for more
information about what voices are available.}

\item{use_gs_pngs}{Use the PNGs that Google will export with
their slide deck.  If \code{FALSE}, must have
\code{animation} package installed.}

\item{...}{Arguments passed to \code{\link{ari_spin}}}
}
\value{
A list of the images, script, and MP4 with the result.
}
\description{
Convert Google Slides and notes to video with ari
}
\note{
The \code{...} argument in \code{gs_tts} corresponds to
arguments passed to \code{\link[text2speech]{tts}}.
}
\examples{
\dontrun{
id = paste0("https://docs.google.com/presentation/d/",
"1Tg-GTGnUPduOtZKYuMoelqUNZnUp3vvg_7TtpUPL7e8",
"/edit#slide=id.g154aa4fae2_0_58")
res = gs_ari(id)
}
}
