% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_convert.R
\name{gs_convert}
\alias{gs_convert}
\title{Convert Google Slides to PNGs and a script}
\usage{
gs_convert(
  id,
  verbose = TRUE,
  PPTX = TRUE,
  auto_stub = TRUE,
  use_gs_ids = FALSE,
  use_gs_pngs = TRUE,
  ...
)
}
\arguments{
\item{id}{File id passed to \code{\link{drive_download}} after passing
through \code{\link{as_id}}}

\item{verbose}{logical, indicating whether to print informative messages,
passed to \code{\link{drive_download}}}

\item{PPTX}{If \code{TRUE}, then the PowerPoint is downloaded as well.
If this is \code{FALSE}, then the script will not be generated.}

\item{auto_stub}{Should the prefix name be taken from
the Google slide name?  If so, `stub` argument will
be overridden in \code{\link{pdf_to_images}}}

\item{use_gs_ids}{use Google slide identifiers for the naming}

\item{use_gs_pngs}{Use the PNGs that Google will export with
their slide deck.  If \code{FALSE}, must have
\code{animation} package installed.}

\item{...}{Additional options to send to \code{\link{pdf_to_images}}
if \code{use_gs_pngs = FALSE}}
}
\value{
A list of the images and the notes for each script
}
\description{
Convert Google Slides to PNGs and a script
}
\examples{
\dontrun{
id = "1XoRj0pwaLI34XKZ7TljVeDHu-tbgGmXRmQa528JIwmw"
PPTX = FALSE
auto_stub = TRUE
res = gs_convert(id, use_gs_ids = TRUE, PPTX = FALSE)
res = gs_convert(id, use_gs_pngs = FALSE)
res = gs_convert(id, use_gs_pngs = TRUE, PPTX = FALSE)
res = gs_convert(id, use_gs_pngs = TRUE, use_gs_ids = TRUE,
PPTX = FALSE)
}
}
