% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanpub_render.R
\name{leanpub_render}
\alias{leanpub_render}
\title{Render Leanpub document}
\usage{
leanpub_render(md_file)
}
\arguments{
\item{md_file}{Path to markdown file}
}
\value{
HTML output filename
}
\description{
Render Leanpub document
}
