% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanpub_render.R
\name{leanpub_render}
\alias{leanpub_render}
\alias{leanpub_document}
\title{Render Leanpub document}
\usage{
leanpub_render(md_file, output_options = list(self_contained = TRUE),
  remove_tests = TRUE, ...)

leanpub_document(..., remove_tests = TRUE, self_contained = TRUE)
}
\arguments{
\item{md_file}{Path to markdown file}

\item{output_options}{arguments to send to \code{\link{render}}}

\item{remove_tests}{Should the quizzes and exercises be removed}

\item{...}{additional arguments to send to \code{\link{render}}
other than \code{input}, `output_dir`, and `output_format`,
and `output_options`}

\item{self_contained}{Produce a standalone HTML file,
passed to \code{\link{html_document}}}
}
\value{
HTML output filename
}
\description{
Render Leanpub document
}
