% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_course.R
\name{translate_course}
\alias{translate_course}
\title{Translate Course to Another Language}
\usage{
translate_course(course_dir, target = "es", detect = TRUE,
  verbose = TRUE, use_book = TRUE, overwrite = FALSE,
  sleep_time = 0, ...)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{target}{language to translate to, see \code{\link{gl_translate}}}

\item{detect}{Should \code{\link{gl_detect_file}} be run to
check that the language is not currently the target?
Must be google Language authorized using \code{\link{gl_auth}}}

\item{verbose}{Print diagnostic messages}

\item{overwrite}{Should files be overwritten?
If not, they are put in a temporary directory and translated.}

\item{sleep_time}{time in seconds to sleep before running
next call}

\item{...}{additional arguments to pass to
\code{\link{translate_manuscript}} and
\code{\link{translate_script}}}
}
\value{
A result from \code{\link{check_course}}
}
\description{
Translate Course to Another Language
}
\examples{
\dontrun{
course_dir = "~/Dropbox/Projects/CDS/cds_dataanalysis"

library(googledrive)
check_didactr_auth()

res = translate_course(course_dir)

}
}
