% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_youtube.R
\name{update_youtube}
\alias{update_youtube}
\title{Update RDS object that tracks YouTube uploads}
\usage{
update_youtube(up = NULL, file = NULL, lesson = NULL,
  metric_path = NULL, save_metrics = TRUE,
  timezone = "America/New_York")
}
\arguments{
\item{up}{output from \code{\link{vids_to_youtube}}}

\item{file}{object created in \code{\link{vids_to_youtube}}}

\item{lesson}{object created in \code{\link{vids_to_youtube}}}

\item{metric_path}{Path to the metrics file}

\item{save_metrics}{Should an `rds` file be saved of the `data.frame`?}

\item{timezone}{Timezone to be used?}
}
\value{
A \code{data.frame} of all uploaded youtube videos
}
\description{
Update RDS object that tracks YouTube uploads
}
