% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_youtube.R
\name{update_youtube}
\alias{update_youtube}
\title{Create PDF and PNGs from Google Slides}
\usage{
update_youtube(up, file = NULL, lesson = NULL, metric_path = NULL)
}
\arguments{
\item{up}{output from \code{\link{vids_to_youtube}}}

\item{file}{object created in \code{\link{vids_to_youtube}}}

\item{lesson}{object created in \code{\link{vids_to_youtube}}}

\item{metric_path}{Path to the metrics file}
}
\value{
A \code{data.frame} of all uploaded youtube videos.
}
\description{
Create PDF and PNGs from Google Slides
}
